refactor: move summary frontend to separate component
