+++ src/App.js
@@ -1,9 +1,5 @@
-import React, { useEffect, useState } from "react";
 import styled from "styled-components";
-import { Title } from "./Components/Title/Title";
-import { Summary } from "./Components/Summary/Summary";
-import useHashChange from "./hooks/useHashChange";
-import extractChat from "./utils/extractChat";
+import { SummaryPage } from "./pages/SummaryPage";
 
 const AppContainer = styled.div`
   width: 400px;
@@ -21,41 +17,9 @@ const AppContainer = styled.div`
 `;
 
 export const App = () => {
-  const [chatInfo, setChatInfo] = useState({});
-  const [loading, setLoading] = useState(false);
-  const [summary, setSummary] = useState("");
-  const [error, setError] = useState("");
-  const hash = useHashChange();
-
-  useEffect(() => {
-    extractChat().then((chatInfo) => setChatInfo(chatInfo));
-  }, [hash])
-
-  useEffect(() => {
-    if (!chatInfo || Object.keys(chatInfo).length === 0) return;
-
-    setLoading(true);
-
-    chrome.runtime.sendMessage({ action: "get_summary", chatInfo }, (response) => {
-      setLoading(false)
-      if (chrome.runtime.lastError) {
-        setError(chrome.runtime.lastError.message);
-        return;
-      }
-      if (response?.error) {
-        setError(response.error);
-        setSummary(null);
-      } else {
-        setError(null);
-        setSummary(response.summary);
-      }
-    });
-    console.log(chatInfo);
-  }, [chatInfo])
   return (
     <AppContainer>
-      <Title />
-      <Summary text={summary} loading={loading} error={error} />
+      <SummaryPage />
     </AppContainer>
   );
 };
